BackEnd/requirements.txt
fastapi
uvicorn
cohere
python-dotenv
tokenizers==0.15.0 